+++ umh-core/pkg/communicator/models/statusmessagev2.go
@@ -12,12 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// This file was generated from JSON Schema using quicktype, do not modify it directly.
-// To parse and unparse this JSON data, add this code to your project and do:
-//
-//    statusMessage, err := UnmarshalStatusMessage(bytes)
-//    bytes, err = statusMessage.Marshal()
-
 package models
 
 import "encoding/json"
@@ -131,9 +125,11 @@ type Connection struct {
 type DFCMetrics struct {
 	FailedMessages float64 `json:"failedMessages"`
 	// Throughput expressed in messages per second
-	ThroughputMsgPerSEC float64 `json:"throughputMsgPerSec"`
-	Unprocessable       float64 `json:"unprocessable"`
-	Unprocessable24H    float64 `json:"unprocessable24h"`
+	ThroughputMsgPerSEC      float64 `json:"throughputMsgPerSec"`
+	ReadThroughputMsgPerSEC  float64 `json:"readThroughputMsgPerSec"`
+	WriteThroughputMsgPerSEC float64 `json:"writeThroughputMsgPerSec"`
+	Unprocessable            float64 `json:"unprocessable"`
+	Unprocessable24H         float64 `json:"unprocessable24h"`
 }
 
 type EventsTable struct {