+++ umh-core/pkg/config/config.go
@@ -72,8 +72,9 @@ type ModelRef struct {
 
 type Field struct {
 	ModelRef     *ModelRef        `yaml:"_refModel,omitempty"`     // this is a special field that is used to reference another data model to be used as a type for this field
+	Relational   *PayloadShape    `yaml:"_relational,omitempty"`   // inline relational data definition with fields and types
 	Subfields    map[string]Field `yaml:",inline"`                 // subfields of the field (allow recursive definition of fields)
-	PayloadShape string           `yaml:"_payloadshape,omitempty"` // type of the field (timeseries only for now)
+	PayloadShape string           `yaml:"_payloadshape,omitempty"` // payload shape reference for timeseries data
 }
 
 type InternalConfig struct {